Import Upstream version 247.3

+++ src/shared/reboot-util.c
@@ -1,4 +1,4 @@
-/* SPDX-License-Identifier: LGPL-2.1+ */
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
 
 #include <errno.h>
 #include <unistd.h>
@@ -6,16 +6,20 @@
 #include "alloc-util.h"
 #include "fileio.h"
 #include "log.h"
+#include "proc-cmdline.h"
 #include "raw-reboot.h"
 #include "reboot-util.h"
 #include "string-util.h"
 #include "umask-util.h"
 #include "virt.h"
 
-int update_reboot_parameter_and_warn(const char *parameter) {
+int update_reboot_parameter_and_warn(const char *parameter, bool keep) {
         int r;
 
         if (isempty(parameter)) {
+                if (keep)
+                        return 0;
+
                 if (unlink("/run/systemd/reboot-param") < 0) {
                         if (errno == ENOENT)
                                 return 0;
@@ -36,15 +40,27 @@ int update_reboot_parameter_and_warn(const char *parameter) {
         return 0;
 }
 
+int read_reboot_parameter(char **parameter) {
+        int r;
+
+        assert(parameter);
+
+        r = read_one_line_file("/run/systemd/reboot-param", parameter);
+        if (r < 0 && r != -ENOENT)
+                return log_debug_errno(r, "Failed to read /run/systemd/reboot-param: %m");
+
+        return 0;
+}
+
 int reboot_with_parameter(RebootFlags flags) {
         int r;
 
-        /* Reboots the system with a parameter that is read from /run/systemd/reboot-param. Returns 0 if REBOOT_DRY_RUN
-         * was set and the actual reboot operation was hence skipped. If REBOOT_FALLBACK is set and the reboot with
-         * parameter doesn't work out a fallback to classic reboot() is attempted. If REBOOT_FALLBACK is not set, 0 is
-         * returned instead, which should be considered indication for the caller to fall back to reboot() on its own,
-         * or somehow else deal with this. If REBOOT_LOG is specified will log about what it is going to do, as well as
-         * all errors. */
+        /* Reboots the system with a parameter that is read from /run/systemd/reboot-param. Returns 0 if
+         * REBOOT_DRY_RUN was set and the actual reboot operation was hence skipped. If REBOOT_FALLBACK is
+         * set and the reboot with parameter doesn't work out a fallback to classic reboot() is attempted. If
+         * REBOOT_FALLBACK is not set, 0 is returned instead, which should be considered indication for the
+         * caller to fall back to reboot() on its own, or somehow else deal with this. If REBOOT_LOG is
+         * specified will log about what it is going to do, as well as all errors. */
 
         if (detect_container() == 0) {
                 _cleanup_free_ char *parameter = NULL;
@@ -55,7 +71,6 @@ int reboot_with_parameter(RebootFlags flags) {
                                        "Failed to read reboot parameter file, ignoring: %m");
 
                 if (!isempty(parameter)) {
-
                         log_full(flags & REBOOT_LOG ? LOG_INFO : LOG_DEBUG,
                                  "Rebooting with argument '%s'.", parameter);
 
@@ -80,4 +95,15 @@ int reboot_with_parameter(RebootFlags flags) {
         (void) reboot(RB_AUTOBOOT);
 
         return log_full_errno(flags & REBOOT_LOG ? LOG_ERR : LOG_DEBUG, errno, "Failed to reboot: %m");
+}
+
+int shall_restore_state(void) {
+        bool ret;
+        int r;
+
+        r = proc_cmdline_get_bool("systemd.restore_state", &ret);
+        if (r < 0)
+                return r;
+
+        return r > 0 ? ret : true;
 }
